Examination module by Shashank

+++ class-section-list.php
@@ -56,7 +56,7 @@ require_once 'includes/sidebar.php';
 											<a href="class-section.php?id=<?php echo $value['id']; ?>" class="btn btn-primary btn-sm mb-1">
 												<i class="fa fa-pencil" aria-hidden="true"></i>
 											</a>
-											<a href="employee/process/processClassSections.php?type=deleteClasses&id=<?php echo $value['id']; ?>" class="btn btn-danger btn-sm mb-1">
+											<a href="employee/process/processClassSections.php?type=deleteClasses&id=<?php echo $value['id']; ?>" onclick="return confirm('Are you sure you want to delete this?')" class="btn btn-danger btn-sm mb-1">
                                                 <i class="fa fa-trash" aria-hidden="true"></i>
                                             </a>
 										</td>
